Move controls documentation do devices.
As the controls are highly dependent on the buttons available on a
specific device hardware it makes sense to document them per device.

To prevent to much redundancy we are using mkdocs-macros to define
includes and variables which allows for a maximum of reuse.
The new structure also makes it possible to only document emulators
actually available on the device.

This initial commit only implements complete documentation for
Powkiddy RGB10 MAX 3, a minimal example is also added to AYN Loki Zero
in order to demonstrate the use of variables.